The executor_interface has a schedule_monitor
