+++ mrs/scheduling/monitor.py
@@ -1,9 +1,5 @@
 import logging
 
-from fmlib.models.tasks import Task
-from ropod.structs.task import TaskStatus as TaskStatusConst
-
-from mrs.execution.interface import ExecutorInterface
 from mrs.scheduling.scheduler import Scheduler
 
 
@@ -20,8 +16,7 @@ class ScheduleMonitor:
     def __init__(self, robot_id,
                  stp_solver,
                  allocation_method,
-                 corrective_measure,
-                 **kwargs):
+                 corrective_measure):
         """ Includes methods to monitor the schedule of a robot's allocated tasks
 
        Args:
@@ -30,24 +25,12 @@ class ScheduleMonitor:
             stp_solver (STP): Simple Temporal Problem object
             allocation_method (str): Name of the allocation method
             corrective_measure (str): Name of the corrective measure
-            kwargs:
-                api (API): object that provides middleware functionality
-                robot_store (robot_store): interface to interact with the db
-
         """
         self.robot_id = robot_id
         self.stp_solver = stp_solver
-        self.api = kwargs.get('api')
-        self.ccu_store = kwargs.get('ccu_store')
-        self.task_queue = None
-
-        self.logger = logging.getLogger('mrs.schedule.monitor.%s' % self.robot_id)
-
         self.corrective_measure = self.get_corrective_measure(allocation_method, corrective_measure)
-
         self.scheduler = Scheduler(self.stp_solver, self.robot_id)
-        self.executor_interface = ExecutorInterface(self.robot_id)
-
+        self.logger = logging.getLogger('mrs.schedule.monitor.%s' % self.robot_id)
         self.logger.debug("ScheduleMonitor initialized %s", self.robot_id)
 
     def get_corrective_measure(self, allocation_method, corrective_measure):
@@ -57,21 +40,4 @@ class ScheduleMonitor:
             raise ValueError(corrective_measure)
 
         return corrective_measure
-
-    def configure(self, **kwargs):
-        api = kwargs.get('api')
-        ccu_store = kwargs.get('ccu_store')
-        if api:
-            self.api = api
-        if ccu_store:
-            self.ccu_store = ccu_store
-
-    def task_cb(self, msg):
-        payload = msg['payload']
-        task = Task.from_payload(payload)
-        task.update_status(TaskStatusConst.DISPATCHED)
-        self.logger.debug("Task %s received", task.task_id)
-        # TODO: Add task to task_queue
-
-
